fix: Tratamento de erros

+++ CrudCSharp/Controllers/ProdutoController.cs
@@ -2,6 +2,7 @@
 using CrudCSharp.Model;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Http;
 
 namespace CrudCSharp.Controllers
 {
@@ -17,22 +18,44 @@ namespace CrudCSharp.Controllers
         }
 
         [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<IEnumerable<Produto>>> GetProdutos() 
         {
+            if (_context == null) return NotFound();
             return await _context.Produtos.ToListAsync();
         }
 
         [HttpGet("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> GetProduto(int id)
         {
+            if (_context == null) return NotFound();
+
             var produto = await _context.Produtos.FindAsync(id);
+
+            if(produto == null) return NotFound();
 
             return Ok(produto);
         }
 
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> PostProduto(Produto produto)
         {
+            if (_context == null) return Problem("Erro ao criar um produto, contate o suporte!");
+
+
+            if(!ModelState.IsValid)
+            {
+                return ValidationProblem(ModelState);
+            }
+
             _context.Produtos.Add(produto);
             await _context.SaveChangesAsync();
 
@@ -40,22 +63,59 @@ namespace CrudCSharp.Controllers
         }
 
         [HttpPut("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> PutProduto(int id, Produto produto)
         {
-            _context.Produtos.Update(produto);
-            await _context.SaveChangesAsync();
+            if (id != produto.Id) return BadRequest();
+
+            if(!ModelState.IsValid) return ValidationProblem(ModelState);
+
+            _context.Entry(produto).State = EntityState.Modified;
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+
+                if (!ProdutoExists(id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+            }
             return NoContent();
         }
 
         [HttpDelete("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(typeof(Microsoft.AspNetCore.Mvc.ModelBinding.ModelStateDictionary), StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<IActionResult> DeleteProduto(int id)
         {
+            if (!ModelState.IsValid) return ValidationProblem(ModelState);
+
             var produto = await _context.Produtos.FindAsync(id);
 
+            if (produto == null) return NotFound();
+
             _context.Produtos.Remove(produto);
-            _context.SaveChangesAsync();
+            await _context.SaveChangesAsync();
 
             return NoContent();
+        }
+
+        private bool ProdutoExists(int id)
+        {
+            return (_context.Produtos?.Any(e => e.Id == id)).GetValueOrDefault();
         }
 
     }
